libdefs: Remove `react-native-webview` libdef, for now.
To reduce the size of the upcoming RN v0.60 upgrade. There, we will
upgrade to the latest 7.x of `react-native-webview`.

There isn't an off-the-shelf libdef from FlowTyped for 7.x, so we
compose our own. That addition is pushed out of the main RN upgrade
commit, just like this one, to reduce the size of the RN upgrade
commit.