deps: Upgrade to latest `react-native-webview` that works with RN v0.59.
Any later (7.0.0 or above), and we'd see peer dependency warnings
about React Native. We'll upgrade to the latest 7.x in the same
commit as the upcoming RN v0.60 upgrade.

AndroidX support is the one declared breaking change [1] for version
6.x.x. We started using AndroidX in e433197cf; if we hadn't, we
wouldn't be able to take this version.

Also update the libdef; FlowTyped conveniently has one for version 6
(but no later):

```
flow-typed install react-native-webview@6
```

[1]: https://github.com/react-native-community/react-native-webview#versioning