deps: Add and use `expo-screen-orientation`.
This is a best approximation of the functionality previously
provided by `react-native-orientation`, which is unmaintained and is
giving us errors that hint it will be unusable in the future.

We followed the installation instructions [1] for bare React Native
apps, which meant changing a line in ios/ZulipMobile/AppDelegate.m.

This note in those instructions appears misleading:

"""
or if you want to change the default screen orientation, with:
"""

Looking at the relevant Apple doc [2], though, it's not to set the
default orientation, but to set supported orientations.  That's also
what we find empirically: if we use the simple version of the code
suggested in those instructions, the effect is actually to lock the
app's orientation to a single value.

Instead we choose to support all orientations, with
`UIInterfaceOrientationMaskAll`.

That setup step is done in the same commit as the JavaScript package
is used; we assume the `AppDelegate.m` changes will break something
if they coincide with the use of `react-native-orientation`.

[1]: https://github.com/expo/expo/tree/ce446ed8b/packages/expo-screen-orientation

[2]: https://developer.apple.com/documentation/uikit/uiinterfaceorientationmask?language=objc