deps: Upgrade React Native from v0.60 to v0.61.
Using the RN Upgrade Helper, a web app showing the diff from the
release/0.60.6 and the release/0.61.5 branches of the
`react-native-community/rn-diff-purge` repo, at
https://react-native-community.github.io/upgrade-helper/?from=0.60.6&to=0.61.5.

In this commit:

- Upgrade `react-native`, `react`, and `flow-bin` following the template

- Change our Podfile to reflect new/changed iOS dependencies in RN,
  following the template

- Make changes to adapt to multiple upgrades of RN's Hermes
  dependency (details below)

- Run `yarn yarn-deduplicate && yarn` as prompted by
  `tools/test deps`

See (on the issue) a list of changes from the upgrade helper that we
don't do in this series [1].

Hermes details:

In facebook/react-native@c21e36db4, React Native started using
v0.1.1 of Hermes [2], which includes a rename of the NPM package
(facebook/hermes@c74842e) from 'hermesvm' to 'hermes-engine'. So,
use the new path in the one place where its path occurs in our code,
following the changes to the template in that commit.

In facebook/react-native@06c64f5f1, React Native started using
v0.2.1 of Hermes [3], which was came with the announcement, "The C++
runtime library is now packaged in a separate AAR to avoid
`pickFirst` nondeterminisms". Follow the changes to the template in
that commit, where several `pickFirst` lines are removed.

[1]: https://github.com/zulip/zulip-mobile/issues/3781#issuecomment-642151537
[2]: https://github.com/facebook/hermes/releases/tag/v0.1.1
[3]: https://github.com/facebook/hermes/releases/tag/v0.2.1

Fixes: #3781