react: Rename unsafe lifecycle methods.
In the upcoming RN v0.60 -> v0.61 upgrade [1], we'll also upgrade to
React v16.9.0, which continues the path toward deprecating some
lifecycle methods by logging warnings to the console unless you
prefix the methods with `UNSAFE_` [2].

Some background on how these methods came to be considered harmful:
https://reactjs.org/blog/2018/03/27/update-on-async-rendering.html.

React v16.9.0 suggests (without requiring) the following [2]:

"""
However, maybe you don’t have the time to migrate or test these
components. In that case, we recommend running a “codemod” script
that renames them automatically:

```
cd your_project
npx react-codemod rename-unsafe-lifecycles
```
"""

So, run that script. That means that we won't get noisy warnings in
the console, but the method names will stand out as requiring a fix
at some point; in particular, in an as-yet-unreleased React v17.

Actually removing our uses of these methods is #4154.

[1]: https://react-native-community.github.io/upgrade-helper/?from=0.60.6&to=0.61.5
[2]: https://reactjs.org/blog/2019/08/08/react-v16.9.0.html