webview: Add support for displaying spoilers.
Support for creating spoilers was recently added to Zulip [1].
Currently, spoilers render as visible by default on the app.

Add functionality to handle spoilers properly - spoilers are
collapsed by default, and can be toggled by tapping on the header.
Also, if a header is missing, the text 'Spoiler' is added as the
header.

The implementation is similar to [1], but the JavaScript was
rewritten to work without jQuery, and the SCSS has been translated
to plain CSS.

[1] https://github.com/zulip/zulip/commit/1cb040647b37f9f2269c2b39ae6495830f49ac4b

Co-authored-by: Chris Bobbe <cbobbe@zulip.com>
Fixes: #4155