jest: Add `jest-expo` preset, to be used in the next commit.
A bit frustratingly, we need to install react-native-web, because
it's a peer dep of jest-expo, and react-dom, because that's a peer
dep of react-native-web.

It's likely that we'll never run any code in either module, since we
don't care about the web. (Expo seems to be expanding React Native's
original claim for multi-platform compatibility, namely iOS and
Android, to include the web. `react-native-web` looks like a thing
that lets you write React Native...for the web.)

At least they don't complain if we add them as dev dependencies
instead of regular dependencies.

Also, run `yarn yarn-deduplicate && yarn` as prompted by
`tools/test deps`.