apiTypes [nfc]: Prevent growth of a cycle-blob of imports.
Soon, we'd like to add an import of src/boot/store.js from
src/utils/logging.js, so we can send the server version to Sentry,
for #3745.

As Greg demonstrates with a handy tool he made to topologically sort
the import graph, that particular import would make an existing
cycle-blob much larger (the technical term is "strongly-connected
component" or SCC) if nothing were done about it. So, do something
about it -- empirically, this commit prevents that growth [1].
Without doing this, and with that import in place, Flow seems to
have a bug: it produces over a megabyte of clearly wrong error
output.

We'd like to keep `ApiResponseServerSettings` (and some aliases it
uses) alongside the code that actually makes the `server_settings`
request [2]; it would be a shame to have to make a separation there
(e.g., by putting the types in src/api/apiTypes.js) just because of
this Flow output.

But, among similar pairs (network-request code with types that
describe the response), this is the only one where we import and
re-export the types through src/api/apiTypes.js. It's not too
awkward just to import directly from the getServerSettings file, and
delete the `export type * from ...` line in src/api/apiTypesjs. So,
do that.

[1] See Greg's report of the improvement at
    https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/import.20cycles/near/1057141.

[2] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M3745.20Server.20version.20to.20Sentry/near/930324