fetchActions tests: Mock response, so `fetchMessages` is error-free.
Without mocking the response [1], errors are thrown from
`fetchMessages` for things like not being able to call `.json()` on
the response.

So, mock the response, like we already do in one test that cares
about the MESSAGE_FETCH_COMPLETE action.

The errors are actually ignored, because we're not awaiting
`store.dispatch(fetchMessages(...))`. We'll fix that in an upcoming
commit.

The assertions that `actions` is of length 1 depend, oddly, on these
errors being thrown. We could assert that the length is two
(MESSAGE_FETCH_START plus MESSAGE_FETCH_COMPLETE) but these tests
only examine the first action.

[1] Here, we mock it to something boring, since we're not examining
    it; the tests are examining features of the MESSAGE_FETCH_START
    action.