accountsReducer: Remove unhelpful test.
With Flow v0.113.0, which we'll get in the RN v0.61 -> v0.62 upgrade
(#3782), red flags are raised about this test. Strangely [1], the
test was allowed to expect the `zulipVersion` on an Account to be
undefined. In fact, the type for it is `ZulipVersion | null`.

The test is based on the false premise that a REALM_INIT action
might not have a `zulipVersion` property. `RealmInitAction` states
that it will always be there, as a ZulipVersion instance. So, delete
the test. If we ever find that there's a legitimate case where it is
missing, we'll fix the RealmInitAction type accordingly, and a new
test can proceed from there.

[1] I think the culprit is very likely facebook/flow#6715, which Ray
    reported as being fixed in Flow v0.111.1, at
    https://github.com/zulip/zulip-mobile/pull/4046#discussion_r412562505