sharing: Don't expect `NativeModules.Sharing` to be defined on iOS.
This avoids a TypeError on startup on iOS, which breaks the
execution of important further setup in AppEventHandlers's
`componentDidMount`.

We may or may not use the same `NativeModules.Sharing` for an iOS
implementation, but I think it's common for RN libraries to
implement Android and iOS modules at the same name in NativeModules.
So this seemed like the most idiomatic way to do this hotfix.