pmConversationsSelectors tests: Fix flakes in two tests.
I forgot that calling `.sort()` on an array of numbers will
not...sort them numerically [1].

Forgetting to pass a custom selector function to do numeric sorting
resulted in two problems, which both occur in both affected tests:

- our definition of the expected results was sometimes incorrect
  (`.ids` in items in the array)

- our faked initial state was sometimes not representative of the
  real state (`state.unread.huddles[].user_ids_string`; see the note
  on `HuddlesUnreadItem` in src/api/initialDataTypes.js)

I say "sometimes" because one of the numbers in the array was
`eg.selfUser.user_id`, which is randomized. The flake was triggered
when string sorting resulted in a different array than numeric
sorting, and this only happened for some values of
`eg.selfUser.user_id` (here, those that started with 1 but had at
least two digits).

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort#Description