tests: Put account in state in several tests that call `fetchMessages`.
It's normal and boring for `fetchMessages` to be dispatched when
there's a logged-in account. So, represent that in the initial
state.

So, continue the work of 38aa238db, and replace the cases we found
in that commit with the slightly preferable `eg.selfAccount`; it's
not the account of just any user, but the user who is logged in and
active (the self).

In the next commit, when `fetchMessages` starts dispatching the
recently added MESSAGE_FETCH_ERROR on errors, we can add tests that
do things like intentionally omit a logged-in account and make sure
Redux is notified.