logo: Fix a version of the logo that shows up pixelated.
We don't use this much -- it's only on the account-pick screen and
only in the case where you've removed all the accounts you'd
previously added.  But when we do, it comes out pixelated.  The
reason is that the image is only 40x40, while we use it at a size of
40x40 logical pixels, which is typically much bigger.  Fix that.