deps: Upgrade @react-native-community/netinfo to latest.
Oddly, when we do a soft upgrade of all our Jest-related
dependencies, we get some errors with stack traces from this package
that mention the Android Support Library. Not sure why.

But it's best to be on the most up-to-date versions of our
dependencies that we can. Often, we're limited in what versions we
can take by compatibility with whatever RN version we're on. This
package has a handy table of compatibility with RN versions (most
packages don't do this!), and I'm glad to see that it doesn't
declare a problem we would have with doing this upgrade.
Furthermore, no peer-dependency warnings appear when we run the
upgrade command.

Declared breaking changes:

- 4.0.0: "android: You are required to only use either the Support
  Library or AndroidX for your dependencies. If you need to migrate
  this library back to the support library, or another library
  forward to AndroidX, then take a look at the [Jetifier
  tool](https://github.com/mikehardy/jetifier)."

  We've been using AndroidX since e433197cf, and we're all set up
  with mikehardy/jetifier to automatically tweak the non-AndroidX
  libraries we use, on `postinstall`, so that they speak AndroidX.

- 5.0.0: "Previously deprecated methods have been removed. These
  methods have been deprecated since this library was extracted from
  the core of React Native. Most users will not have any issues with
  migrating if they were not ignoring the previous warnings."

  I just checked the doc for the new version [1], and we're not
  using anything we're not supposed to be; the one thing we access
  on the default export of the module is `.addEventListener`, and we
  use it correctly. We also haven't seen any deprecation warnings.

[1] https://github.com/react-native-community/react-native-netinfo/blob/v5.9.5/README.md#api