messageActions tests: Refocus test coverage onto isFetchNeededAtAnchor.
Two tests for `doNarrow` implicitly test the behavior of
`isFetchNeededAtAnchor`:

- 'when no messages in new narrow but caught up, only switch to
  narrow, do not fetch'

- 'when messages in new narrow are too few and not caught up, fetch
  more messages'

It might have been better if `isFetchNeededAtAnchor` had had these
focused tests to begin with. But we're about to have `doNarrow`
never do a fetch, so it'll be quite unreasonable for them to still
cover `isFetchNeededAtAnchor`.

So, remove the tests, and add focused tests for
`isFetchNeededAtAnchor`. Another thing that both removed tested was
that a navigation action is dispatched with `doNarrow`. But this is
already covered in the first `doNarrow` test.