eventTypes [nfc]: Synchronize EventTypes class with `eventToAction`.
It looks like we're using this class to store string constants for
the `type` property the server uses to label events. We haven't used
the constants in as many places as we could, but might as well bring
them up-to-date.

It now contains all the event types that we handle in
`eventToAction`.