eventMiddleware: Rename; run on actions instead of events.
Following from discussion [1].

In a recent commit, we took out the parts of `eventMiddleware` that
mutate event objects, and we started doing the same crunchy-shell
conversions in `eventToAction` instead, and we had them not mutate
anything.

But `eventToAction` still takes care of running some side effects
that don't mutate the inputs, like playing a sound. Playing a sound
is the only thing it currently does, but it's plausible that we'll
want a central place like this to do more side effects, as long as
they don't mutate the inputs. (In fact, we'll soon move the only
call site of `clearTypingNotification` into here.)

We also take a further step to strengthen `eventToAction`'s claim to
be the crunchy shell [2] for data we get from events: we bring
`eventToAction` closer to the edge by having it run before the code
in `eventMiddleware`, and we rename `eventMiddleware` because it's
not in the middle anymore (and it was too vaguely named anyway). The
function now takes an EventAction as an input instead of an event.
An EventAction should be considered more trustworthy than an untyped
event object.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/user_avatar_url_field_optional/near/965896
[2] https://github.com/zulip/zulip-mobile/blob/master/docs/architecture/crunchy-shell.md