crunchy shell: Use AvatarURL for Message objects.
And have the UI logic that handles avatars for messages accept the
new form.

We'll do the same for UserOrBot objects in the next commit.

There's a small amount of churn that comes from not converting users
and messages in the same commit; I've added temporary code comments
where this happens.

For the `GET /messages` response, the natural place to do the
conversion is in `migrateMessages`. For events, it's in
`eventToAction`.