types [nfc]: Move some abstract type utilities to their own file.
We're about to add another utility type of the same flavor as these,
so it's time to make them an appropriate home.

Our `types.js` is primarily about the specific types used in the
app, which doesn't make the most logical home for these.  We've had
some in the file `react-redux.js` because that's where they're used,
but they're really much more general and it's nice to have a home
for them that's convenient for other modules to import from.