jest libdef: Add `getRealSystemTime` and `setSystemTime`.
These are reflected in the doc [1] [2] and in the `Jest` interface
in `node_modules/@jest/environment/build/index.d.ts`.

A small difference in the comments: the doc says "modern fake timers
implementation", where the TypeScript says "Lolex as fake timers
implementation". I've sided with the docs, as it's probably wise to
treat Lolex as an implementation detail of "modern" fake timers that
may change.

Adjust the entry under `[libs]` as Ray suggests in chat [3].

[1] https://jestjs.io/docs/en/jest-object#jestgetrealsystemtime
[2] https://jestjs.io/docs/en/jest-object#jestsetsystemtimenow-number--date
[3] https://chat.zulip.org/#narrow/stream/48-mobile/topic/flow-typed.20for.20Jest/near/854562