events [nfc]: Perform side effects before dispatching, vs. after.
The one side effect we currently do here is indifferent to whether
it happens before or after we dispatch the event actions.  But we
have another side effect found in the oddly-named `actionCreator`
function, which we'd like to fold into this mechanism, and for that
one it's essential that it happens before the events are dispatched.
So, reorder to accommodate that.

If in the future we have some other side effect that needs to happen
after dispatch, and still have this one or another that needs to
happen before, then we can have both kinds separately.