avatar: Pipe `Account[0].realm` as URL down to `getAvatarUrl`.
As in the previous and next commits, stringify the realm exactly
where we need it as a string.

This slightly expands the area of our codebase in which it's easy to
see that the realm is well-formed data.

`getAvatarUrl` itself will likely disappear soon, when we make the
shell crunchier for handling `.avatar_url` on users, cross-realm
bots, and messages.

But the changes we make to its callers, here, will continue to be
beneficial after that.