getServerSettings: Pipe realm as URL down.
As in the previous and next commits, stringify the realm exactly
where we need it as a string.

This slightly expands the area of our codebase in which it's easy to
see that the realm is well-formed data.

Also rename `this.state.realm` to `this.state.realmInputValue`, to
be more descriptive.