deps: Upgrade rollup-related dependencies.
Two of these have moved to the `@rollup`-scoped NPM packages; one of
these moves was hinted at in the previous commit.

We had to change the imports slightly; it looks like
@rollup/plugin-babel and @rollup/plugin-node-resolve use CommonJS
modules.

We rename `rollupResolve` to the more descriptive
`rollupNodeResolve`, which, like `rollupBabel`, answers a desire to
match the example in the corresponding doc [1] [2] while recognizing
that the variables share a namespace with everything else in
`tools/generate-webview-js`; we don't use a separate
`rollup.config.js` file, as the examples do.

[1] https://github.com/rollup/plugins/tree/master/packages/babel
[2] https://github.com/rollup/plugins/tree/master/packages/node-resolve