android: Prevent activity recreation on theme change.
Part of the RN v0.61 -> v0.62 changes to the template app [1],
corresponding to facebook/react-native@83a16b16c. This is a bugfix
that we're free to do at any time as long as we don't have any code
that depends on the "UI mode" but neglects to listen for changes to
it. We don't currently need use the UI mode at all, and we won't
consider it until #4009, for which the API isn't available until the
upcoming RN v0.62 upgrade.

See more detail from Greg on GitHub [2].

[1] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2
[2] https://github.com/zulip/zulip-mobile/pull/4244#discussion_r482424989