ios build: Address two Xcode deprecation warnings.
Part of the RN v0.61 -> v0.62 changes to the template app [1],
corresponding to facebook/react-native@ebb629d05. These are warnings
I think we've seen for a while (#4112), and there's no suggestion
that we can't do them before the main upgrade commit.

The diff in the RN commit seems to address two warnings at the same
time, even though only the "Missing Localizability" one is
mentioned; like in that commit, we do them both here.

They are:

"""
Migrate "English.lproj" (Deprecated)

Migrating the English, deprecated localization to English is
recommended for all projects. This will ensure localized resources
are placed in “en.lproj” directories instead of deprecated
“English.lproj” directories.
"""

"""
Turn on "Missing Localizability"

This will turn on the static analyzer check for "Missing
Localizability", because this project is localized for multiple
languages.
"""

The "Missing Localizability" change appears in the Upgrade Support
repo's official guide for how to change the Xcode files using
Xcode [2]; the "Migrate 'english.lproj' (Deprecated)" one does not.

See also discussion on GitHub [3].

[1] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2
[2] https://github.com/react-native-community/upgrade-support/issues/13
[3] https://github.com/zulip/zulip-mobile/pull/4244#discussion_r482566498