Input [nfc]: Remove unused `this.textInput`.
This was only used in `this.handleClear`; our use of that was
deleted in b0a02fa48. With that swept away in the current commit, we
can remove `this.textInput`; so, do.

Now, `Input` itself doesn't need to use the ref of the TextInput
component; only the users of `Input` may need to do that. Therefore,
we don't need to fall back on the callback style for refs [1]; we
can use the more modern `createRef` [2] which we'll do in an
upcoming commit.

[1] https://reactjs.org/docs/refs-and-the-dom.html#callback-refs
[2] https://reactjs.org/docs/refs-and-the-dom.html#creating-refs