Touchable [nfc]: Inline value for `background` prop on Android.
After the upcoming RN v0.61 -> v0.62 upgrade, Flow would correctly
flag the fact that we're trying to compare a `number | string` value
to a number. (`Platform.Version` is a number on Android and a string
on iOS [1].)

With the new placement of the `Platform.Version` check, Flow will be
able to infer that `Platform.Version` is a number from the
conditional on `Platform.OS` above. (It looks like RN types
`Platform` as a union of types corresponding to each platform, so
`Platform.OS` and `Platform.Version` correctly correspond with each
other.)

[1] https://reactnative.dev/docs/platform-specific-code#detecting-the-android-version