transport types: Make `getAuthHeaders` return type more precise.
It's not a mystery which auth headers we might want to get from this
function; there's no reason we have to use `{ [string]: string }`
here.

We just have to do something a bit silly in the fix, like we did in
d141be763, and make a note of the corresponding Flow issue in a
comment.