flow: Point to new location of a RN-internal file.
Part of the RN v0.61 -> v0.62 changes to the template app [1],
corresponding to the insertion in the Flow config in
facebook/react-native@e54ecf907.

After the main upgrade commit, we'll finish following
facebook/react-native@e54ecf907 by making its two deletions in the
Flow config. The deleted lines point to the old, then-incorrect
location of the file.

[1] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2