flow [nfc]: Adjust comments about when a Flow change happens.
See discussion [1], where Greg points out that it looks like a
mistake to suppose that this change would come with v0.111; there's
nothing at that version in the changelog [2] that looks related.

He points out that there *is* something related at the entry for
v0.126.0:

"""
* Allow indexers in exact object annotations
"""

And that the minimal example provided works at v0.126.0, but not
before [3]:

```
const val: {| +[string]: number |} = { foo: 3 };
```

There isn't currently a version of React Native out that uses v0.126
or later, unfortunately; v0.63 is the latest, and it uses v0.122.
But react-native's `master` branch is on v0.133, which is actually
the latest!

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Flow.20v0.2E113.20upgrade/near/1012737
[2] https://github.com/facebook/flow/blob/master/Changelog.md#01110
[3] https://flow.org/try/#0MYewdgzgLgBAbgQwDYC4YG8A+MDa0BOAlmAOYC6aYArgLYBGApvjJgL4wC8GMAZiCGgDMMVgG4AUEA