msglist css [nfc]: Group CSS rules by broad area of what they do.
Hopefully this will help in reading these styles and seeing all
the rules that interact in a given area of the message list.

By putting all the rules aimed at message content in one place,
it should also help with attempting to compare to the webapp's
rendered_markdown.css to see how our styles have diverged.

To keep this change tractable to read and review (at least with the
help of `git diff --color-moved` or equivalent), we stick purely to
reordering whole CSS rules, and adding a few blank lines and comments
outside the rules.  There are plenty of opportunities for further
cleanup in this file.

So long as I haven't made a scribal error by inadvertently changing
something beyond the order of rules (and comments and whitespace),
I believe there's just one way this change could unintentionally
have an effect on behavior: if it swaps the order of some rules
that can apply to the same element, have the same precedence, and
say contrary things.  To try to prevent that, I've generally
preserved the order of groups of rules that are aimed at the same
part of the message list.  But that seems the most likely form of
bug that may be lurking.