deps: Upgrade react-redux to v6.
v7 is the latest, but might as well get there incrementally.

The v6 release seems most excited to talk about its new use of the
new Context API, which sounds like it might mean a lot of breaking
changes we have to address manually (I guess I'm thinking of #4222).
But no; in fact, there's just one small announced breaking change
that applies to us:

- The `withRef` option to `connect` has been replaced with
  `forwardRef`. If `{forwardRef : true}` has been passed to
  `connect`, adding a ref to the connected wrapper component will
  actually return the instance of the wrapped component.

So, make that change, including the implied removal of
`getWrappedInstance`; that bit of UI works fine on Android and iOS
after that removal.

There's a FlowTyped libdef for this version, so, take that. An
additional Flow error arose, falsely telling consumers of our
`connect`-wrapped `MentionWarnings` component that they need to pass
props like `auth`; in fact, those are not expected to be passed
because they're provided by `connect`. Just as we've done at the
`connect` call site, add a temporary $FlowFixMe until we can get
`connect` properly type-checked.

It looks like we don't want to linger on v6 for very long; v7
addresses some performance complaints that arose in v6. So, we'll
move to v7 ASAP.