RealmScreen: Remove last use of `isValidUrl`.
And use a URL computation instead, which is what we do elsewhere. In
particular, it's how we check whether we can store the realm in
Redux, so it's good to make the UI align with that. This helper
function is a relic from the days before #4081 and #4146, so it
should disappear anyway.

It was the sole user of an NPM package. So, uninstall that.