ChatScreen: Get `isFocused` from React Navigation.
As noted in discussion [1], we haven't been using quite the right
information to decide whether it's time for a ChatScreen to fetch
new messages. As Greg says, there,

"""
[T]he condition we'd like to have governing this isn't "is this
ChatScreen showing the same narrow that's topmost or last", but
rather "is this the topmost or last ChatScreen".
"""

To grab that bit, we first considered [2] something like adding a
unique ID to each ChatScreen, which seems hacky.

But while I was doing the recent React Navigation v3 and v4
upgrades, I came across this `withNavigationFocus` HOC, which
provides an `isFocused` prop [3], which looks more or less exactly
like what we'd want. "Is this ChatScreen focused" isn't *exactly*
the same question as "is this the topmost or last ChatScreen", but
the differences seem trivial, and I think this commit's approach is
simpler and clearer. See also Greg's thoughts on the difference [4].

There's also an `isFocused` *method* directly on the `navigation`
prop [5]; we get the `navigation` prop for free [6] since we've
stuck `ChatScreen` directly on one of our navigators. The problem
with just using that method is that `ChatScreen` doesn't naturally
update (re-`render`) when it goes in and out of focus, so our code
in `componentDidUpdate` that would ask for the `isFocused` value
would never get run.

So, use this HOC, which does set up `ChatScreen` to listen for a
change of focus.

[1] https://github.com/zulip/zulip-mobile/pull/4205#pullrequestreview-461268523
[2] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4156.20Message.20List.20placeholders/near/978256
[3] https://reactnavigation.org/docs/4.x/function-after-focusing-screen#triggering-an-action-with-the-withnavigationfocus-higher-order-component
[4] https://github.com/zulip/zulip-mobile/pull/4269#pullrequestreview-498729733
[5] https://reactnavigation.org/docs/4.x/navigation-prop/#isfocused---query-the-focused-state-of-the-screen
[6] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4156.20Message.20List.20placeholders/near/953303