store: Spell out `getNav` here, before we reimplement `getNav`.
As in the previous few commits, this is one of a small handful of
places where we need to write code to explicitly get the navigation
state from Redux, before we reimplement `getNav` to get the state
from the recently added `NavigationService`.

Like in those previous few commits, this is a place where we'll no
longer need the navigation state, when #3804 is completed.