tabs [nfc]: Remove two options that don't do anything for us.
We use the default export from this file (calling it `tabsOptions`
in all cases`) for both `createBottomTabNavigator` and
`createMaterialTopTabNavigator`.

Both of these two options were actually removed from
`createBottomTabNavigator`'s config in react-navigation v2 [1], with
much controversy [2]. So neither of them has been doing anything for
`MainTabs`, our only use of `createBottomTabNavigator`, for quite a
while. When we start using a proper libdef in an upcoming commit, we
would get errors about passing these two options to
`createBottomTabNavigator`.

The `swipeEnabled` option remains available for
`createMaterialTopTabNavigator`. We have three uses of
`createMaterialTopTabNavigator`, and they all have `swipeEnabled`
set to `true` (we did this in ad0b5f3e0) so we don't even have to
think about what the default value is, and this is NFC for those.

`animationEnabled` isn't documented for
`createMaterialTopTabNavigator`, so its removal should also be NFC.
It gets a mention in the v3 doc [3] but not the v4 doc [4].

[1] https://reactnavigation.org/blog/2018/05/07/react-navigation-2.0#tab-navigator-split-into-separate-components
[2] https://github.com/react-navigation/react-navigation/issues/4146
[3] https://reactnavigation.org/docs/3.x/material-top-tab-navigator
[4] https://reactnavigation.org/docs/4.x/material-top-tab-navigator