tabs: Separate "material-top" and "bottom" tab-nav config helpers.
There are differences in the config options between
`createBottomTabNavigator` and `createMaterialTopTabNavigator`,
which Flow will soon start telling us about, when we start using an
actual libdef.

So, be specific about which options we give which type of navigator.

Testing on iOS and Android shows that this doesn't obviously changed
anything. It shouldn't, anyway, because the status quo is that we've
been handing `createBottomTabNavigator` some options that (if the
docs are to be trusted) it has no idea how to handle, and plausibly
just drops on the floor.

And in particular, the removal of the #2065 workaround
(`borderWidth: 0` and `elevation: 0`, targeting material top tabs)
from the bottom-tab-nav config doesn't seem to give any adverse
effects.