navActions: Let `navigateToRealmScreen` take `initial`.
Our current nav-state logic has a case where `RealmScreen` has
`params.initial` in its navigation state. So, allow this to be
passed in the action creator.

The case is a weird one: it came about in 7a04a5638, when
`getStateForRoute` started including `initial: true` when the passed
`route` was 'realm'. This has coincided with the fact that
`getStateForRoute`'s only call site with 'realm' for `route` has
been one where it's appropriate to have `initial: true`.

Anyway, that call site is in the `rehydrate` handler in
`navReducer`, which we're going to dissolve quite soon, using the
`navigateToRealmScreen` action creator instead.